重新整理既有javascript
