--- app/javascript/games/app.js
@@ -1,10 +1,0 @@
-import { Game } from "games/game"
-
-export function run(app){
-  if ( $('#maincontainer').length ){
-    app = new Game();
-  } else if ( app && app.timerID ){
-    clearInterval( app.timerID );
-  }
-  return app;
-}
