+++ app/javascript/games/index.js
@@ -1,4 +1,14 @@
-import { run } from "games/app"
-import { hidden_alert } from "games/sessions"
+import { Game } from "./game"
+import { hidden_alert } from "./sessions"
 
-export { run, hidden_alert }
+export { hidden_alert }
+
+export function run(app){
+  if ( $('#maincontainer').length ){
+    app = new Game();
+    app.request_status();
+  } else if ( app && app.timerID ){
+    clearInterval( app.timerID );
+  }
+  return app;
+}
